Conflicts:
	hphp/test/run

+++ hphp/runtime/ext/ext_thread.cpp
@@ -60,8 +60,8 @@ int64_t f_hphp_get_thread_id() {
   return  (unsigned long)Process::GetThreadId();
 }
 
-int32_t f_hphp_gettid() {
-    return (unsigned int)Process::GetThreadPid();
+int64_t f_hphp_gettid() {
+  return (unsigned int)Process::GetThreadPid();
 }
 
 ///////////////////////////////////////////////////////////////////////////////
