+++ hphp/runtime/ext/ext_thread.h
@@ -31,7 +31,7 @@ bool f_hphp_service_thread_stopped(int timeout);
 bool f_hphp_thread_is_warmup_enabled();
 void f_hphp_thread_set_warmup_enabled();
 int64_t f_hphp_get_thread_id();
-int32_t f_hphp_gettid();
+int64_t f_hphp_gettid();
 
 ///////////////////////////////////////////////////////////////////////////////
 }
